+++ R/data.R
@@ -4,6 +4,7 @@
 #' @description Data on IGOs from 1815-2014, at the IGO-year level.
 #' Contains one record per IGO-year (with years listed at 5 year intervals
 #' through 1965, and annually thereafter).
+
 #' @source \href{https://correlatesofwar.org/data-sets/IGOs}{Intergovernmental Organizations (v3)},
 #' The Correlates of War Project (IGO Data Stata Files)
 #' @format data frame with 19335 rows. Relevant fields:
@@ -38,7 +39,7 @@
 #' Pevehouse, Jon CW, Timothy Nordstrom, Roseanne W McManus, and Anne Spencer
 #' Jamison. "Tracking Organizations in the World: The Correlates of War
 #' IGO Version 3.0 Datasets." \emph{Journal of Peace Research} 57, no. 3
-#' (May 2020): 492-503. \url{https://doi.org/10.1177/0022343319881175}.
+#' (May 2020): 492-503. \doi{10.1177/0022343319881175}.
 NULL
 
 #' @title Country membership to IGO by year
@@ -52,7 +53,7 @@ NULL
 #' @format data frame with 15557 rows. Relevant fields:
 #' \itemize{
 #'   \item{\strong{ccode}: }{COW country number,
-#'   see \code{\link{cow_country_codes}}}.
+#'   see \code{\link{states2016}}}.
 #'   \item{\strong{year}: }{Calendar Year.}
 #'   \item{\strong{state}: }{Abbreviated state name, identical to variable
 #'   names in \code{\link{igo_year_format3}}}
@@ -72,35 +73,62 @@ NULL
 #'
 #' See \href{https://correlatesofwar.org/data-sets/IGOs}{\strong{Codebook Version 3 IGO Data}}
 #' @seealso \code{\link{igo_year_format3}},
-#' \code{\link{cow_country_codes}}. See also \code{\link[countrycode]{countrycode}} to
+#' \code{\link{states2016}}. See also \code{\link[countrycode]{countrycode}} to
 #' convert between different country code schemes.
 #' @note Raw data used internally by \pkg{igoR}
 #' @references
 #' Pevehouse, Jon CW, Timothy Nordstrom, Roseanne W McManus, and Anne Spencer
 #' Jamison. "Tracking Organizations in the World: The Correlates of War
 #' IGO Version 3.0 Datasets." \emph{Journal of Peace Research} 57, no. 3
-#' (May 2020): 492-503. \url{https://doi.org/10.1177/0022343319881175}.
+#' (May 2020): 492-503. \doi{10.1177/0022343319881175}.
 NULL
 
 
-#' @title COW Country Codes
-#' @name cow_country_codes
+#' @title State System Membership (v2016)
+#' @name states2016
 #' @docType data
 #' @description The list of states with COW abbreviations and ID numbers, plus
 #' the field \code{state} from \code{\link{state_year_format3}}.
-#' @source \href{https://correlatesofwar.org/data-sets/cow-country-codes}{
-#' COW Country Codes}, The Correlates of War Project.
-#' @format data frame with 217 rows:
+#' @source \href{https://correlatesofwar.org/data-sets/state-system-membership}{
+#' State System Membership (v2016)}, The Correlates of War Project.
+#' @format data frame with 243 rows:
 #' \itemize{
 #'   \item{\strong{ccode}: }{COW country number.}
 #'   \item{\strong{stateabb}: }{COW state abbreviation (3 characters).}
 #'   \item{\strong{statenme}: }{COW state name.}
+#'  \item{\strong{styear...endday}: }{Fields to identify the begining
+#'   and the end of each tenure.}
+#'   \item{\strong{version}: }{Data file version number.}
 #'   \item{\strong{state}: }{Abbreviated state name as appear in
 #'   \code{\link{state_year_format3}}.}
 #'  }
 #'
 #' @seealso \code{\link{state_year_format3}}, \code{\link{igo_year_format3}}.
+#' @details This data set contains the list of states in the international
+#' system as updated and distributed by the Correlates of War Project.
+#'
+#' This data set contains the list of states in the international system as
+#' updated and distributed by the Correlates of War Project.
+#' These data sets identify states, their standard Correlates of War
+#' "country code" or state number (used throughout the Correlates of War
+#' project data sets), state abbreviations, and dates of membership as
+#' states and major powers in the international system.
+#'
+#' The Correlates of War project includes a state in the international system
+#' from 1816-2016 for the following criteria:
+#'
+#' \strong{Prior to 1920} the entity must have had a population greater than 500,000
+#' and have had diplomatic missions at or above the rank of charge
+#' d'affaires with Britain and France.
+#'
+#' \strong{After 1920} the entity must be a member of the League of Nations
+#' or the United Nations, or have a population greater than 500,000 and
+#' receive diplomatic missions from two major powers.
+#'
 #' @note \code{state} variable added to original data to help comparison across
 #' datasets on this package.
-NULL
 #'
+#' @references
+#' Correlates of War Project. 2017 "State System Membership List, v2016."
+#' Online, \url{http://correlatesofwar.org}.
+NULL