+++ R/igoR-package.R
@@ -12,7 +12,8 @@
 #' @name igoR-package
 #' @aliases igoR
 #' @docType package
-#' @description \if{html}{\figure{logo.png}{options: width=120 alt="igoR logo" align='right'}}
+#' @description
+#' \if{html}{\figure{logo.png}{options: width=120 alt="igoR" align='right'}}
 #' Tools to extract information from the Intergovernmental
 #' Organizations Database ('IGO'), version 3, provided by the Correlates of
 #' War Project \url{https://correlatesofwar.org/}.
@@ -27,7 +28,7 @@
 #' Pevehouse, Jon CW, Timothy Nordstrom, Roseanne W McManus, and Anne Spencer
 #' Jamison. "Tracking Organizations in the World: The Correlates of War
 #' IGO Version 3.0 Datasets." \emph{Journal of Peace Research} 57, no. 3
-#' (May 2020): 492-503. \url{https://doi.org/10.1177/0022343319881175}.
+#' (May 2020): 492-503. \doi{10.1177/0022343319881175}.
 #'
 #' @seealso
 #' Useful links:
@@ -40,5 +41,5 @@
 NULL
 
 # import stuffs
-#' @importFrom stats aggregate
+#' @importFrom stats aggregate reshape
 NULL