+++ NEWS.md
@@ -1,3 +1,14 @@
+# igoR 0.1.0
+
+* Vectorize `igo_members()`
+* Generalize search ignoring case.
+* Add `igo_state_membership()`
+* `cow_country_codes` now is internal. 
+* Added `igoR::states2016` in replacement of `cow_country_codes`.
+* Add `igo_dyadic()`
+* Add `igo_search_states()`
+
+
 # igoR 0.0.0.9000
 
 * Pre-release